// backend/migrations/snapshots/f38d4b1766e4994c5919b1005aea4941cc14628b9b7b2e5a29b05fe42e1f0ce4/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Channel: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdBy: CodecTypes['pg/text@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly guildId: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedBy: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly Conversation: {
      readonly channelId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdBy: CodecTypes['pg/text@1']['output'] | null;
      readonly directKey: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly type: 'CHANNEL' | 'DIRECT' | 'AI';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedBy: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly ConversationMember: {
      readonly conversationId: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly Guild: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdBy: CodecTypes['pg/text@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly ownerId: CodecTypes['pg/uuid@1']['output'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedBy: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly GuildMember: {
      readonly guildId: CodecTypes['pg/uuid@1']['output'];
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly Message: {
      readonly authorId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly clientMessageId: CodecTypes['pg/uuid@1']['output'] | null;
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly conversationId: CodecTypes['pg/uuid@1']['output'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdBy: CodecTypes['pg/text@1']['output'] | null;
      readonly editedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly revokeAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly senderType: 'USER' | 'ASSISTANT';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedBy: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly createdBy: CodecTypes['pg/text@1']['output'] | null;
      readonly deletedAt: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly role: 'MEMBER' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updatedBy: CodecTypes['pg/text@1']['output'] | null;
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Channel: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdBy: CodecTypes['pg/text@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly guildId: CodecTypes['pg/uuid@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedBy: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly Conversation: {
      readonly channelId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdBy: CodecTypes['pg/text@1']['input'] | null;
      readonly directKey: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly type: 'CHANNEL' | 'DIRECT' | 'AI';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedBy: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly ConversationMember: {
      readonly conversationId: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly Guild: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdBy: CodecTypes['pg/text@1']['input'] | null;
      readonly description: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly name: CodecTypes['pg/text@1']['input'];
      readonly ownerId: CodecTypes['pg/uuid@1']['input'];
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedBy: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly GuildMember: {
      readonly guildId: CodecTypes['pg/uuid@1']['input'];
      readonly joinedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly userId: CodecTypes['pg/uuid@1']['input'];
    };
    readonly Message: {
      readonly authorId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly clientMessageId: CodecTypes['pg/uuid@1']['input'] | null;
      readonly content: CodecTypes['pg/text@1']['input'];
      readonly conversationId: CodecTypes['pg/uuid@1']['input'];
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdBy: CodecTypes['pg/text@1']['input'] | null;
      readonly editedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly revokeAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly senderType: 'USER' | 'ASSISTANT';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedBy: CodecTypes['pg/text@1']['input'] | null;
    };
    readonly User: {
      readonly createdAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly createdBy: CodecTypes['pg/text@1']['input'] | null;
      readonly deletedAt: CodecTypes['pg/timestamptz-string@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/uuid@1']['input'];
      readonly passwordHash: CodecTypes['pg/text@1']['input'];
      readonly role: 'MEMBER' | 'ADMIN';
      readonly updatedAt: CodecTypes['pg/timestamptz-string@1']['input'];
      readonly updatedBy: CodecTypes['pg/text@1']['input'] | null;
      readonly username: CodecTypes['pg/text@1']['input'] | null;
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly channels: {
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly created_by: CodecTypes['pg/text@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly guildId: CodecTypes['pg/uuid@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updated_by: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly conversation_members: {
      readonly conversationId: CodecTypes['pg/uuid@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly conversations: {
      readonly channel_id: CodecTypes['pg/uuid@1']['output'] | null;
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly created_by: CodecTypes['pg/text@1']['output'] | null;
      readonly direct_key: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly type: 'CHANNEL' | 'DIRECT' | 'AI';
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updated_by: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly guild_members: {
      readonly guildId: CodecTypes['pg/uuid@1']['output'];
      readonly joined_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly userId: CodecTypes['pg/uuid@1']['output'];
    };
    readonly guilds: {
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly created_by: CodecTypes['pg/text@1']['output'] | null;
      readonly description: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly name: CodecTypes['pg/text@1']['output'];
      readonly ownerId: CodecTypes['pg/uuid@1']['output'];
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updated_by: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly messages: {
      readonly author_id: CodecTypes['pg/uuid@1']['output'] | null;
      readonly client_message_id: CodecTypes['pg/uuid@1']['output'] | null;
      readonly content: CodecTypes['pg/text@1']['output'];
      readonly conversation_id: CodecTypes['pg/uuid@1']['output'];
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly created_by: CodecTypes['pg/text@1']['output'] | null;
      readonly edited_at: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly revoke_at: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly sender_type: 'USER' | 'ASSISTANT';
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updated_by: CodecTypes['pg/text@1']['output'] | null;
    };
    readonly users: {
      readonly created_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly created_by: CodecTypes['pg/text@1']['output'] | null;
      readonly deleted_at: CodecTypes['pg/timestamptz-string@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/uuid@1']['output'];
      readonly passwordHash: CodecTypes['pg/text@1']['output'];
      readonly role: 'MEMBER' | 'ADMIN';
      readonly updated_at: CodecTypes['pg/timestamptz-string@1']['output'];
      readonly updated_by: CodecTypes['pg/text@1']['output'] | null;
      readonly username: CodecTypes['pg/text@1']['output'] | null;
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;